Change sitemap generation
pulling in all the static version api docs was too much for our current AI solution to ingest, so I'm paring back to a manually defined single version for now.

Manual definition intends to track the most recent version that is actually released, so AI can pull this content to reasonably answer all but the most specific version questions.